[infra] deploy lambda with SAM CLI
Previously to deploy a lambda I would have to `pip install` the dependencies into a folder locally, zip the folder up, then `terraform apply` to deploy the lambda. The main problem with this approach is that my development machine is not on the same pip platform as Lambda, so wheels are not compatible. Another problem is the maximum package size of 50 MB when deploying via AWS API (not S3).

To avoid these problems, I've refactored the `create_db_user_with_iam_role` Lambda to use the AWS Serverless Application Model (SAM). This commit adds a script that uses the SAM CLI to install dependencies in a docker container that matches the Lambda execution environment, then uploads the package to S3 for deployment by Terraform.